LPS-54791 Bookmarks: Standardize locations for internal setting classes
