+++ modules/apps/bookmarks/bookmarks-api/src/com/liferay/bookmarks/settings/internal/BookmarksGroupServiceSettingsIdMapping.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.bookmarks.settings.definition;
+package com.liferay.bookmarks.settings.internal;
 
 import com.liferay.bookmarks.configuration.BookmarksGroupServiceConfiguration;
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.bookmarks.constants.BookmarksConstants;
+import com.liferay.portal.kernel.settings.definition.SettingsIdMapping;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Iván Zaera
  */
 @Component
-public class BookmarksGroupServiceConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
+public class BookmarksGroupServiceSettingsIdMapping
+	implements SettingsIdMapping {
 
 	@Override
 	public Class getConfigurationBeanClass() {
 		return BookmarksGroupServiceConfiguration.class;
+	}
+
+	@Override
+	public String getSettingsId() {
+		return BookmarksConstants.SERVICE_NAME;
 	}
 
 }
